refactor: share OpenID4VP request resolution
